+++ lector_facturas_app/lib/providers/scan_list_provider.dart
@@ -4,7 +4,7 @@ import 'package:lector_facturas_app/providers/db_provider.dart';
 class ScanListProvider extends ChangeNotifier{
   List<ScanModel> scans = [];
 
-  Future<ScanModel> nuevoScan(String cufe,String fecha, int total, String numFactura, String doc, String establecimiento ) async {
+  Future<ScanModel> nuevoScan(String cufe,String fecha, double total, String numFactura, String doc, String establecimiento ) async {
     final nuevoScan = new ScanModel(cufe: cufe,fecha: fecha,total: total,num_factura: numFactura,doc: doc,establecimiento: establecimiento);
     final id = await DBProvider.db.nuevoScanRaw(nuevoScan);
     nuevoScan.id = id;
